requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-telegram-bot
rapidfuzz
groq
GitPython
python-docx
pydantic
